+++ rustybeer-cli/src/main.rs
@@ -6,7 +6,7 @@ fn main() {
     let app = App::new("RustyBeer")
         .version("0.1")
         .subcommand(commands::abv::add_subcommand())
-        .subcommand(commands::abv_abw::add_subcommand())
+        .subcommand(commands::alcohol_volume_weight::add_subcommand())
         .subcommand(commands::beer_style::add_subcommand())
         .subcommand(commands::boil_off::add_subcommand())
         .subcommand(commands::diluting::add_subcommand())
@@ -22,7 +22,7 @@ fn main() {
                 commands::abv::do_matches(&app);
             }
             "abv_abw" => {
-                commands::abv_abw::do_matches(&app);
+                commands::alcohol_volume_weight::do_matches(&app);
             }
             "beer_style" => {
                 commands::beer_style::do_matches(&app);
