+++ rustybeer/src/calculators/mod.rs
@@ -1,5 +1,5 @@
 pub mod abv;
-pub mod abv_abw;
+pub mod alcohol_volume_weight;
 pub mod diluting;
 pub mod ibu;
 pub mod num_bottles;
